Mipmapping and minor refactoring
Texture formats now specify only the minification filter, and magnification is always set to GL_LINEAR, so mipmapping filters can be used.
Automatic generation of mipmaps will be attempted when loading textures.

Modified Entity to use the Vec3f class for vector data, instead of a float-array.
Added support for specifying a model to an entity, and drawing the entity.